changed menu selection method to numbers

+++ ATM.Tests/ATMTests.cs
@@ -11,10 +11,10 @@ namespace ATM.Tests
         //public void Does_balance_exist()
         //{
         //    //Arrange
-        //    double balance = 20;
+        //    decimal balance = 20;
 
         //    // Act
-        //    double result = Lab02UnitTests.ATM.ViewBalance(balance);
+        //    decimal result = Lab02UnitTests.ATM.ViewBalance(balance);
 
         //    // Assert
         //    Assert.StrictEqual(balance, result);
@@ -25,12 +25,12 @@ namespace ATM.Tests
         public void  Does_money_come_out()
         {
             //Arrange
-            double balance = Lab02UnitTests.ATM.GetBalance();
-            double withdrawal = 10;
-            double testBalance = balance - withdrawal;
+            decimal balance = Lab02UnitTests.ATM.GetBalance();
+            decimal withdrawal = 10;
+            decimal testBalance = balance - withdrawal;
 
             // Act
-            double result = Lab02UnitTests.ATM.WithdrawFunds(withdrawal);
+            decimal result = Lab02UnitTests.ATM.WithdrawFunds(withdrawal);
 
             // Assert
             Assert.Equal(testBalance, result);
@@ -40,12 +40,12 @@ namespace ATM.Tests
         public void Does_negative_money_come_out()
         {
             //Arrange
-            double balance = Lab02UnitTests.ATM.GetBalance();
-            double withdrawal = -10;
-            double testBalance = balance - withdrawal;
+            decimal balance = Lab02UnitTests.ATM.GetBalance();
+            decimal withdrawal = -10;
+            decimal testBalance = balance - withdrawal;
 
             // Act
-            double result = Lab02UnitTests.ATM.WithdrawFunds(withdrawal);
+            decimal result = Lab02UnitTests.ATM.WithdrawFunds(withdrawal);
 
             // Assert
             Assert.Equal(testBalance, result);
@@ -55,14 +55,14 @@ namespace ATM.Tests
         public void Does_money_go_in()
         {
             //Arrange | Given
-            double balance = Lab02UnitTests.ATM.GetBalance();
-            double deposit = 10;
+            decimal balance = Lab02UnitTests.ATM.GetBalance();
+            decimal deposit = 10;
 
             // Act | When
-            double result = Lab02UnitTests.ATM.DepositFunds(deposit);
+            decimal result = Lab02UnitTests.ATM.DepositFunds(deposit);
 
             // Assert | Then
-            double expectedBalance = balance + deposit;
+            decimal expectedBalance = balance + deposit;
             Assert.Equal(expectedBalance, result);
         }
 
@@ -70,14 +70,14 @@ namespace ATM.Tests
         public void Does_negative_money_go_in()
         {
             //Arrange
-            double balance = Lab02UnitTests.ATM.GetBalance();
-            double deposit = -10;
+            decimal balance = Lab02UnitTests.ATM.GetBalance();
+            decimal deposit = -10;
 
             // Act
-            double result = Lab02UnitTests.ATM.DepositFunds(deposit);
+            decimal result = Lab02UnitTests.ATM.DepositFunds(deposit);
 
             // Assert
-            double expectedBalance = balance + deposit;
+            decimal expectedBalance = balance + deposit;
             Assert.Equal(expectedBalance, result);
         }
 
@@ -85,11 +85,11 @@ namespace ATM.Tests
         public void Overdraft_protection()
         {
             //Arrange
-            double balance = Lab02UnitTests.ATM.GetBalance();
-            double withdrawal = balance + 500;
+            decimal balance = Lab02UnitTests.ATM.GetBalance();
+            decimal withdrawal = balance + 500;
 
             // Act
-            double balanceAfterWithdraw = Lab02UnitTests.ATM.WithdrawFunds(withdrawal);
+            decimal balanceAfterWithdraw = Lab02UnitTests.ATM.WithdrawFunds(withdrawal);
 
 
             // Assert
